[FSTORE-1131] DeltaStreamer job fails when feature group has complex feature (#1178)
* init

* fix
