+++ java/spark/src/main/java/com/logicalclocks/hsfs/spark/engine/hudi/DeltaStreamerKafkaSource.java
@@ -31,8 +31,6 @@ import org.apache.hudi.utilities.sources.AvroSource;
 import org.apache.hudi.utilities.sources.InputBatch;
 import org.apache.hudi.utilities.sources.helpers.KafkaOffsetGen;
 
-import org.apache.kafka.common.header.Header;
-import org.apache.kafka.common.header.Headers;
 import org.apache.kafka.common.serialization.StringDeserializer;
 
 import org.apache.log4j.LogManager;
@@ -44,8 +42,6 @@ import org.apache.spark.sql.SparkSession;
 import org.apache.spark.streaming.kafka010.KafkaUtils;
 import org.apache.spark.streaming.kafka010.LocationStrategies;
 import org.apache.spark.streaming.kafka010.OffsetRange;
-
-import java.nio.charset.StandardCharsets;
 
 public class DeltaStreamerKafkaSource extends AvroSource {
   private static final Logger LOG = LogManager.getLogger(DeltaStreamerKafkaSource.class);
@@ -91,24 +87,16 @@ public class DeltaStreamerKafkaSource extends AvroSource {
     if (totalNewMsgs <= 0L) {
       return new InputBatch(Option.empty(), KafkaOffsetGen.CheckpointUtils.offsetsToStr(offsetRanges));
     } else {
-      JavaRDD<GenericRecord> newDataRdd = this.toRdd(offsetRanges, props.getString(HudiEngine.SUBJECT_ID));
+      JavaRDD<GenericRecord> newDataRdd = this.toRdd(offsetRanges);
       return new InputBatch(Option.of(newDataRdd), KafkaOffsetGen.CheckpointUtils.offsetsToStr(offsetRanges));
     }
   }
 
-  private JavaRDD<GenericRecord> toRdd(OffsetRange[] offsetRanges, String subjectId) {
+  private JavaRDD<GenericRecord> toRdd(OffsetRange[] offsetRanges) {
     return KafkaUtils.createRDD(this.sparkContext, this.offsetGen.getKafkaParams(), offsetRanges,
         LocationStrategies.PreferConsistent())
-            .filter(obj -> subjectId.equals(getHeader(obj.headers(), "subjectId")))
+            .filter(obj -> obj.value() != null)
             .map(obj -> (GenericRecord) obj.value());
-  }
-
-  private static String getHeader(Headers headers, String headerKey) {
-    Header header = headers.lastHeader(headerKey);
-    if (header != null) {
-      return new String(header.value(), StandardCharsets.UTF_8);
-    }
-    return null;
   }
 
   @Override
